default cltv_expiry for invoices 144 blocks
Currently the default cltv expiry of the node is used for invoices. If the
default cltv expiry is too low, however, this increases force close risk with a
signer that is not always online. Increase the default final cltv expiry to 144
blocks, so there's enough leeway for a client to come back online after a period
of time and settle pending htlcs.