swap: wrap invoices for channel open
If the swap invoice is not found locally, but the preimage already exists on the
node, that means we created the invoice on the node before. However, if the
invoice is used for a channel open, the invoice on the node will have a lower
amount than the one that should be sent to the sender. In the case of the swap,
the invoice has to match the sender amount in case of a channel open, so we wrap
the node's invoice if it already exists.