set lsp id and pubkey in the same function
This update ensures that lsp pubkey and lsp id are
always set in the same call. Sometimes the lsp
pubkey is not yet available. Then only the lsp id
will be set. If the lsp id changes, the pubkey is
then removed. If the lsp id remains the same, the
node pubkey will still be persisted.