detect and act on network change
When there is a network change, tonic behaves as follows:
- After the keepalive timeout, reconnect automatically.
- Before the keepalive timeout, any grpc call will time out. After the
timeout, the connection is reestablished.

This commit adds a mechanism to reconnect all grpc clients after one of
the clients detects a network change. Initially it was attempted to only
retry based on a keepalive timeout error, but a network change affects
all grpc clients, so subsequent requests to other grpc endpoints would
still fail with a timeout. Ofcourse those grpc clients can also add a
retry-on-timeout, but since it is known at this point the grpc clients
are temporarily dead, reconnect them immediately. This ensures
subsequent calls to other endpoints won't add additional time by waiting
for a timeout.

tokio was bumped to 1.41 to allow cloning watch::Sender.