ensure pseudo payments are deleted on sync
Before sending out a payment, the sdk inserts a pseudo payment. This
ensures there is always a pending payment shown on clients directly when
paying. If the call to send_payment never resulted in in actual payment
being sent, these payments would sit in the database forever and never
be deleted. Pseudo payments are now marked as pseudo in the database. On
sync pseudo payments are deleted before inserting the updated payments
from the node.

Existing payments and sync state are dropped in case the client's
database already contained pending payments because of this issue.